Using nextElementSibling instead of nextSibling
Using nextElementSibling instead of nextSibling when checking for and
removing the sham hidden input.

This is necessary nextSibling can return a text node for whitespaces in
certain browsers (see
https://developer.mozilla.org/en-US/docs/Web/API/Node/nextSibling), so
target.nextElementSibling.getAttribute in hasSham is undefined and an
exception is thrown.

Please note nextSibling is still used when adding the sham hidden
input, as we want it to sit right next to the element to make readonly.